LayoutTests/fast/js/kde/encode_decode_uri-expected.txt

FAIL checkWithFunctions(encodeURI,decodeURI) should be true. Was false
FAIL checkWithFunctions(encodeURIComponent,decodeURIComponent) should be true. Was false
PASS successfullyParsed is true
